Started porting tests
