+++ core/js/src/test/scala/fs2/TestStrategy.scala
@@ -1,0 +1,3 @@
+package fs2
+
+object TestStrategy
