+++ core/js/src/test/scala/fs2/TestUtilPlatform.scala
@@ -1,0 +1,4 @@
+package fs2
+
+trait TestUtilPlatform {
+}
