+++ core/jvm/src/main/scala/fs2/compress.scala
