+++ core/jvm/src/test/scala/fs2/EventuallySupport.scala
@@ -1,0 +1,7 @@
+package fs2
+
+import org.scalatest.concurrent.Eventually
+
+trait EventuallySupport extends Eventually { self: Fs2Spec =>
+  override implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = timeLimit)
+}
