+++ core/jvm/src/test/scala/fs2/TestStrategy.scala
@@ -1,0 +1,6 @@
+package fs2
+
+object TestStrategy {
+  implicit val S: Strategy = Strategy.fromFixedDaemonPool(8)
+  implicit val scheduler: Scheduler = Scheduler.fromFixedDaemonPool(2)
+}
