+++ core/shared/src/test/scala/fs2/CompilationTest.scala
