+++ core/shared/src/test/scala/fs2/Fs2Spec.scala
@@ -1,20 +1,22 @@
 package fs2
 
-import org.scalatest.{ Args, FreeSpec, Matchers, Status }
-import org.scalatest.concurrent.{ Eventually, TimeLimitedTests }
+import org.scalatest.{ Args, AsyncFreeSpec, FreeSpec, Matchers, Status, Suite }
+import org.scalatest.concurrent.{ AsyncTimeLimitedTests, TimeLimitedTests }
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalatest.time.SpanSugar._
 
-abstract class Fs2Spec extends FreeSpec
+abstract class Fs2Spec extends FreeSpec with Fs2SpecLike with TimeLimitedTests {
+  val timeLimit = 90.seconds
+}
+
+abstract class AsyncFs2Spec extends AsyncFreeSpec with Fs2SpecLike with AsyncTimeLimitedTests {
+  val timeLimit = 90.seconds
+}
+
+trait Fs2SpecLike extends Suite
   with GeneratorDrivenPropertyChecks
   with Matchers
-  with TimeLimitedTests
-  with Eventually
   with TestUtil {
-
-  val timeLimit = 90.seconds
-
-  override implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = timeLimit)
 
   implicit override val generatorDrivenConfig: PropertyCheckConfiguration =
     PropertyCheckConfiguration(minSuccessful = 25, workers = 1)
