+++ core/shared/src/test/scala/fs2/TextSpec.scala
